integrate backwards unit tests
